Speed, I am speed. One winner, forty-two losers. I eat losers for breakfast. Breakfast... maybe I should have had breakfast.
According to all known laws of aviation, there is no way a bee should be able to fly. It's wings are too small to get its fat little body off the ground. The bee, of course, flies anyway, because bees don't care what humans think is impossible.

Once upon a time, there was a lovely princess. But she had an enchantment upon her of a fearful sort, which could only be broken by love's first kiss. She was locked away in a castle guarded by a terrible fire-breathing dragon.